feat: Refactor templates and components structure, migrating Create and Edit forms to new paths
