+++ pages/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import MainLayout from '@/components/layouts/main-layout';
-import Home from '@/components/templatesa/home';
+import Home from '@/components/templates/home';
 
 export default function DashboardPage() {
   return (
