+++ pages/signed-out.tsx
@@ -1,15 +1,4 @@
-import { GetServerSideProps } from "next";
-
-export const getServerSideProps: GetServerSideProps = async (context) => {
-  // Set headers to prevent caching
-  context.res.setHeader('Cache-Control', 'no-store, max-age=0, must-revalidate');
-
-  return {
-    props: {}, // will be passed to the page component as props
-  };
-};
-
-const SignedOutPage = () => {
+export const SignedOutPage = () => {
   return (
     <div className="flex flex-col items-center justify-center min-h-screen">
       <h1 className="text-2xl font-bold">You have been signed out.</h1>
@@ -18,5 +7,3 @@ const SignedOutPage = () => {
     </div>
   );
 };
-
-export default SignedOutPage;
